requirements.txt
Django
djangorestframework
whitenoise